requirements.txt
pyspark==3.5.0
delta-spark==3.2.0
requests
papermill
ipykernel
pandas
apache-airflow-providers-databricks>=6.0.0
dbt-core>=1.8.0
dbt-databricks>=1.8.0